pcap::sendqueue

Module windows

WinPcap/npcap sendqueue support module.

Sending individual packets through WinPcap/npcap can be stunningly slow, since a user-to-kernel transition is required for each packet transfer. To alleviate this there’s support for queueing up a batch of packets in userland, requiring only a single transition to kernel to transmit them all.

Structs

SendQueue

Representation of a batch of packets that can be transferred in a single call using SendQueue::transmit().

Enums

SendSync

Indicate whether to send packets as quickly as possible or delay the relative amount of time between packet header timestamps between packet transmissions.