pcap

Struct BreakLoop

pub struct BreakLoop { /* private fields */ }

BreakLoop can safely be sent to other threads such as signal handlers to abort blocking capture loops such as Capture::next_packet and Capture::for_each.

See https://www.tcpdump.org/manpages/pcap_breakloop.3pcap.html for per-platform caveats about how breakloop can wake up blocked threads.

Implementations

impl BreakLoop

pub fn breakloop(&self)

Calls pcap_breakloop to make the blocking loop of a pcap capture return. The call is a no-op if the handle is invalid.

Safety

Can be called from any thread, but must not be used inside a signal handler unless the owning Capture is guaranteed to still be alive.

The signal handler should defer the execution of BreakLoop::breakloop() to a thread instead for safety.